+++ src/web/ArkRouteErrorHandlerInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace sinri\ark\web;
+
+
+interface ArkRouteErrorHandlerInterface
+{
+    public function execute($errorData = [], $http_code = 404);
+
+}